split tests into folders and extracted some to files
