+++ papiea-engine/__tests__/auth_tests/casbin_authorizer.test.ts
@@ -1,15 +1,15 @@
 import "jest"
-import { UserAuthInfo } from "../src/auth/authn";
-import { PermissionDeniedError, Action, ReadAction, CreateAction, UpdateAction, DeleteAction } from "../src/auth/authz";
-import { CasbinAuthorizer } from "../src/auth/casbin";
+import { UserAuthInfo } from "../../src/auth/authn";
+import { PermissionDeniedError, Action, ReadAction, CreateAction, UpdateAction, DeleteAction } from "../../src/auth/authz";
+import { CasbinAuthorizer } from "../../src/auth/casbin";
 import { resolve } from "path";
 import { readFileSync } from "fs";
 
 
 describe("Casbin authorizer tests", () => {
-    const pathToModel: string = resolve(__dirname, "../src/auth/provider_model_example.txt");
+    const pathToModel: string = resolve(__dirname, "../../src/auth/provider_model_example.txt");
     const modelText: string = readFileSync(pathToModel).toString();
-    const pathToPolicy: string = resolve(__dirname, "../src/auth/provider_policy_example.txt");
+    const pathToPolicy: string = resolve(__dirname, "../../src/auth/provider_policy_example.txt");
     const policyText: string = readFileSync(pathToPolicy).toString();
     const authorizer = new CasbinAuthorizer(modelText, policyText);
     
@@ -168,9 +168,9 @@ describe("Casbin authorizer tests", () => {
 });
 
 describe("Casbin authorizer tests for default provider policy", () => {
-    const pathToModel: string = resolve(__dirname, "../src/auth/provider_model_example.txt");
+    const pathToModel: string = resolve(__dirname, "../../src/auth/provider_model_example.txt");
     const modelText: string = readFileSync(pathToModel).toString();
-    const pathToPolicy: string = resolve(__dirname, "../src/auth/provider_policy_example.txt");
+    const pathToPolicy: string = resolve(__dirname, "../../src/auth/provider_policy_example.txt");
     const policyText: string = readFileSync(pathToPolicy).toString();
     const authorizer = new CasbinAuthorizer(modelText, policyText);
