+++ papiea-engine/__tests__/auth_tests/entity_api_auth.test.ts
@@ -3,7 +3,7 @@ import axios from "axios";
 import * as http from "http";
 const url = require("url");
 const queryString = require("query-string");
-import { ProviderBuilder } from "./test_data_factory";
+import { ProviderBuilder } from "../test_data_factory";
 import uuid = require("uuid");
 import { Metadata, Spec, Provider } from "papiea-core";
