+++ papiea-engine/__tests__/database_tests/database_mongo.test.ts
@@ -1,11 +1,11 @@
 import "jest"
-import { MongoConnection } from "../src/databases/mongo";
-import { Spec_DB } from "../src/databases/spec_db_interface";
-import { Status_DB } from "../src/databases/status_db_interface";
-import { Provider_DB } from "../src/databases/provider_db_interface";
-import { S2S_Key_DB } from "../src/databases/s2skey_db_interface";
+import { MongoConnection } from "../../src/databases/mongo";
+import { Spec_DB } from "../../src/databases/spec_db_interface";
+import { Status_DB } from "../../src/databases/status_db_interface";
+import { Provider_DB } from "../../src/databases/provider_db_interface";
+import { S2S_Key_DB } from "../../src/databases/s2skey_db_interface";
 import { v4 as uuid4 } from 'uuid';
-import { ConflictingEntityError } from "../src/databases/utils/errors";
+import { ConflictingEntityError } from "../../src/databases/utils/errors";
 import { Metadata, Spec, Entity_Reference, Status, Kind, Provider, S2S_Key } from "papiea-core";
 
 declare var process: {
