+++ papiea-engine/__tests__/intentful_engine_tests/sfs.test.ts
@@ -1,6 +1,6 @@
 import "jest"
 
-const papi_clj = require("../papiea-lib-clj/papiea-lib-clj.js").papiea_lib_clj;
+const papi_clj = require("../../papiea-lib-clj/papiea-lib-clj.js").papiea_lib_clj;
 const clj_str = (a: any) => papi_clj.core.clj_str(a);
 const sfs_parser = (sfs_ast: string) => papi_clj.core.parse_sfs(sfs_ast);
 const sfs_optimizer = (sfs_ast: string) => papi_clj.core.optimize_sfs_ast(sfs_ast);
