+++ papiea-engine/__tests__/procedure_tests/procedures_api.test.ts
@@ -1,7 +1,7 @@
 import "jest"
 import * as http from "http"
 import axios from "axios"
-import { ProviderBuilder } from "./test_data_factory"
+import { ProviderBuilder } from "../test_data_factory"
 import { Provider } from "papiea-core"
 
 declare var process: {
