+++ papiea-engine/__tests__/sdk_tests/provider_sdk.test.ts
@@ -3,7 +3,7 @@ import { load } from "js-yaml";
 import { resolve } from "path";
 import { Kind_Builder, ProviderSdk } from "papiea-sdk";
 import { plural } from "pluralize"
-import { loadYaml } from "./test_data_factory";
+import { loadYaml } from "../test_data_factory";
 import axios from "axios"
 import { readFileSync } from "fs";
 import { Procedural_Signature, Procedural_Execution_Strategy } from "papiea-core";
@@ -30,7 +30,7 @@ describe("Provider Sdk tests", () => {
         done();
     });
     const provider_version = "0.1.0";
-    const location_yaml = load(readFileSync(resolve(__dirname, "./location_kind_test_data.yml"), "utf-8"));
+    const location_yaml = load(readFileSync(resolve(__dirname, "../test_data/location_kind_test_data.yml"), "utf-8"));
     test("Yaml parses into walkable tree", (done) => {
         expect(location_yaml).not.toBeNull();
         expect(location_yaml.Location).not.toBeNull();
@@ -147,7 +147,7 @@ describe("Provider Sdk tests", () => {
         const location = sdk.new_kind(location_yaml);
         sdk.version(provider_version);
         sdk.prefix("location_provider");
-        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./procedure_move_input.yml"), loadYaml("./location_kind_test_data.yml"), async (ctx, entity, input) => {
+        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./test_data/procedure_move_input.yml"), loadYaml("./test_data/location_kind_test_data.yml"), async (ctx, entity, input) => {
             entity.spec.x += input;
             const res = await axios.put(ctx.url_for(entity), {
                 spec: entity.spec,
@@ -168,7 +168,7 @@ describe("Provider Sdk tests", () => {
             const location = sdk.new_kind(location_yaml);
             sdk.version(provider_version);
             sdk.prefix("location_provider");
-            location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./procedure_move_input.yml"), loadYaml("./location_kind_test_data.yml"), async (ctx, entity, input) => {
+            location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./test_data/procedure_move_input.yml"), loadYaml("./test_data/location_kind_test_data.yml"), async (ctx, entity, input) => {
                 entity.spec.x += input;
                 const res = await axios.put(ctx.url_for(entity), {
                     spec: entity.spec,
@@ -199,7 +199,7 @@ describe("Provider Sdk tests", () => {
         const location = sdk.new_kind(location_yaml);
         sdk.version(provider_version);
         sdk.prefix("location_provider");
-        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./procedure_move_input.yml"), loadYaml("./location_kind_test_data.yml"), async (ctx, entity, input) => {
+        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./test_data/procedure_move_input.yml"), loadYaml("./test_data/location_kind_test_data.yml"), async (ctx, entity, input) => {
 
             throw new Error("Malformed provider")
 
@@ -229,7 +229,7 @@ describe("Provider Sdk tests", () => {
         const location = sdk.new_kind(location_yaml);
         sdk.version(provider_version);
         try {
-            location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./procedure_move_input.yml"), loadYaml("./location_kind_test_data.yml"), async (ctx, entity, input) => {
+            location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./test_data/procedure_move_input.yml"), loadYaml("./test_data/location_kind_test_data.yml"), async (ctx, entity, input) => {
                 entity.spec.x += input;
                 const res = await axios.put(ctx.url_for(entity), {
                     spec: entity.spec,
@@ -247,7 +247,7 @@ describe("Provider Sdk tests", () => {
         const location = sdk.new_kind(location_yaml);
         sdk.version(provider_version);
         sdk.prefix("location_provider");
-        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./procedure_move_input.yml"), loadYaml("./location_kind_test_data.yml"), async (ctx, entity, input) => {
+        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./test_data/procedure_move_input.yml"), loadYaml("./test_data/location_kind_test_data.yml"), async (ctx, entity, input) => {
             entity.spec.x += input;
             const res = await axios.put(ctx.url_for(entity), {
                 spec: entity.spec,
@@ -258,8 +258,8 @@ describe("Provider Sdk tests", () => {
         location.kind_procedure(
             "computeGeolocation",
             {}, Procedural_Execution_Strategy.Halt_Intentful,
-            loadYaml("./procedure_geolocation_compute_input.yml"),
-            loadYaml("./procedure_geolocation_compute_input.yml"), async (ctx, input) => {
+            loadYaml("./test_data/procedure_geolocation_compute_input.yml"),
+            loadYaml("./test_data/procedure_geolocation_compute_input.yml"), async (ctx, input) => {
                 let cluster_location = "us.west.";
                 cluster_location += input;
                 return cluster_location
@@ -279,7 +279,7 @@ describe("Provider Sdk tests", () => {
         const location = sdk.new_kind(location_yaml);
         sdk.version(provider_version);
         sdk.prefix("location_provider");
-        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./procedure_move_input.yml"), loadYaml("./location_kind_test_data.yml"), async (ctx, entity, input) => {
+        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./test_data/procedure_move_input.yml"), loadYaml("./test_data/location_kind_test_data.yml"), async (ctx, entity, input) => {
             entity.spec.x += input;
             const res = await axios.put(ctx.url_for(entity), {
                 spec: entity.spec,
@@ -290,8 +290,8 @@ describe("Provider Sdk tests", () => {
         location.kind_procedure(
             "computeGeolocation",
             {}, Procedural_Execution_Strategy.Halt_Intentful,
-            loadYaml("./procedure_geolocation_compute_input.yml"),
-            loadYaml("./procedure_geolocation_compute_input.yml"), async (ctx, input) => {
+            loadYaml("./test_data/procedure_geolocation_compute_input.yml"),
+            loadYaml("./test_data/procedure_geolocation_compute_input.yml"), async (ctx, input) => {
                 let cluster_location = "us.west.";
                 cluster_location += input;
                 return cluster_location
@@ -312,7 +312,7 @@ describe("Provider Sdk tests", () => {
         const location = sdk.new_kind(location_yaml);
         sdk.version(provider_version);
         sdk.prefix("location_provider");
-        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./procedure_move_input.yml"), loadYaml("./location_kind_test_data.yml"), async (ctx, entity, input) => {
+        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./test_data/procedure_move_input.yml"), loadYaml("./test_data/location_kind_test_data.yml"), async (ctx, entity, input) => {
             entity.spec.x += input;
             const res = await axios.put(ctx.url_for(entity), {
                 spec: entity.spec,
@@ -323,8 +323,8 @@ describe("Provider Sdk tests", () => {
         sdk.provider_procedure("computeSum",
             {},
             Procedural_Execution_Strategy.Halt_Intentful,
-            loadYaml("./procedure_sum_input.yml"),
-            loadYaml("./procedure_sum_output.yml"),
+            loadYaml("./test_data/procedure_sum_input.yml"),
+            loadYaml("./test_data/procedure_sum_output.yml"),
             async (ctx, input) => {
                 return input.a + input.b;
             }
@@ -342,7 +342,7 @@ describe("Provider Sdk tests", () => {
         const location = sdk.new_kind(location_yaml);
         sdk.version(provider_version);
         sdk.prefix("location_provider");
-        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./procedure_move_input.yml"), loadYaml("./location_kind_test_data.yml"), async (ctx, entity, input) => {
+        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./test_data/procedure_move_input.yml"), loadYaml("./test_data/location_kind_test_data.yml"), async (ctx, entity, input) => {
             entity.spec.x += input;
             const res = await axios.put(ctx.url_for(entity), {
                 spec: entity.spec,
@@ -353,8 +353,8 @@ describe("Provider Sdk tests", () => {
         sdk.provider_procedure("computeSum",
             {},
             Procedural_Execution_Strategy.Halt_Intentful,
-            loadYaml("./procedure_sum_input.yml"),
-            loadYaml("./procedure_sum_output.yml"),
+            loadYaml("./test_data/procedure_sum_input.yml"),
+            loadYaml("./test_data/procedure_sum_output.yml"),
             async (ctx, input) => {
                 return input.a + input.b;
             }
@@ -374,7 +374,7 @@ describe("Provider Sdk tests", () => {
         const location = sdk.new_kind(location_yaml);
         sdk.version(provider_version);
         sdk.prefix("location_provider");
-        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./procedure_move_input.yml"), loadYaml("./location_kind_test_data.yml"), async (ctx, entity, input) => {
+        location.entity_procedure("moveX", {}, Procedural_Execution_Strategy.Halt_Intentful, loadYaml("./test_data/procedure_move_input.yml"), loadYaml("./test_data/location_kind_test_data.yml"), async (ctx, entity, input) => {
             entity.spec.x += input;
             const res = await axios.put(ctx.url_for(entity), {
                 spec: entity.spec,
@@ -385,8 +385,8 @@ describe("Provider Sdk tests", () => {
         sdk.provider_procedure("computeSum",
             {},
             Procedural_Execution_Strategy.Halt_Intentful,
-            loadYaml("./procedure_sum_input.yml"),
-            loadYaml("./procedure_sum_output.yml"),
+            loadYaml("./test_data/procedure_sum_input.yml"),
+            loadYaml("./test_data/procedure_sum_output.yml"),
             async (ctx, input) => {
                 return "Totally not a number should fail provider-level validation";
             }
@@ -457,8 +457,8 @@ describe("Provider Sdk tests", () => {
         sdk.provider_procedure("computeSumThrowsError",
             {},
             Procedural_Execution_Strategy.Halt_Intentful,
-            loadYaml("./procedure_sum_input.yml"),
-            loadYaml("./procedure_sum_output.yml"),
+            loadYaml("./test_data/procedure_sum_input.yml"),
+            loadYaml("./test_data/procedure_sum_output.yml"),
             async (ctx, input) => {
                 throw new Error("My custom error")
             }
