+++ papiea-engine/jest.setup.js
@@ -1,0 +1,1 @@
+jest.setTimeout(5000);
